+++ src/client/client-log.ts
@@ -1,5 +1,8 @@
+import { BUILD } from "@app-data";
 
-export const STENCIL_DEV_MODE = ['%cstencil', 'color: white;background:#4c47ff;font-weight: bold; font-size:10px; padding:2px 6px; border-radius: 5px'];
+export const STENCIL_DEV_MODE = BUILD.isTesting
+  ? ['STENCIL:'] // E2E testing
+  : ['%cstencil', 'color: white;background:#4c47ff;font-weight: bold; font-size:10px; padding:2px 6px; border-radius: 5px'];
 
 export const consoleDevError = (...m: any[]) => console.error(...STENCIL_DEV_MODE, ...m);