+++ src/compiler_next/config/validate-config.ts
@@ -55,6 +55,7 @@ export const validateConfig = (userConfig?: Config) => {
   setBooleanConfig(config, 'minifyCss', null, !config.devMode);
   setBooleanConfig(config, 'minifyJs', null, !config.devMode);
   setBooleanConfig(config, 'buildEs5', 'es5', !config.devMode);
+  setBooleanConfig(config, 'buildDocs', 'docs', !config.devMode);
   setBooleanConfig(config, 'writeLog', 'log', false);
   setBooleanConfig(config, 'buildAppCore', null, true);
   setBooleanConfig(config, 'autoprefixCss', null, config.buildEs5);