+++ src/declarations/stencil-private.ts
@@ -26,6 +26,7 @@ import {
   PrerenderRequest,
   PrerenderResults,
   StencilSystem,
+  StyleDoc,
 } from './stencil-public-compiler';
 
 import {
@@ -2251,6 +2252,7 @@ export interface TransformCssToEsmInput {
   commentOriginalSelector?: boolean;
   sourceMap?: boolean;
   minify?: boolean;
+  docs?: boolean;
   autoprefixer?: any;
 }
 
@@ -2260,6 +2262,7 @@ export interface TransformCssToEsmOutput {
   map: any;
   diagnostics: Diagnostic[];
   defaultVarName: string;
+  styleDocs: StyleDoc[];
   imports: { varName: string; importPath: string;}[];
 }