+++ src/declarations/stencil-public-runtime.ts
@@ -425,9 +425,7 @@ export interface ComponentInterface {
   [memberName: string]: any;
 }
 
-/**
- * General types important to applications using stencil built components
- */
+// General types important to applications using stencil built components
 export interface EventEmitter<T= any> {
   emit: (data?: T) => CustomEvent<T>;
 }
@@ -461,7 +459,7 @@ export interface FunctionalUtilities {
 }
 
 export interface FunctionalComponent<T = {}> {
-  (props: T, children: VNode[], utils: FunctionalUtilities): VNode | VNode[];
+  (props: T | undefined, children: VNode[], utils: FunctionalUtilities): VNode | VNode[];
 }
 
 export interface ChildNode {