+++ src/testing/mock-fetch.ts
@@ -1,3 +1,4 @@
+import { MockHeaders, MockRequestInfo, MockResponse } from '@stencil/core/mock-doc';
 import { URL } from 'url';
 
 
@@ -30,11 +31,11 @@ async function globalMockFetch(requestInput: MockRequestInfo) {
     throw new Error(`invalid url for mock fetch()`);
   }
 
-  requestUrl = (new URL(requestUrl, FETCH_BASE_URL)).href;
+  requestUrl = (new URL(requestUrl, location.href)).href;
 
   let mockedData = mockedResponses.get(requestUrl);
   if (mockedData == null) {
-    const defaultUrl = new URL(FETCH_DEFAULT_PATH, FETCH_BASE_URL);
+    const defaultUrl = new URL(FETCH_DEFAULT_PATH, location.href);
     mockedData = mockedResponses.get(defaultUrl.href);
   }
 
@@ -80,7 +81,7 @@ function setMockedResponse(response: MockResponse, input: MockRequestInfo, rejec
   if (!response) {
     throw new Error('MockResponse required');
   }
-  if (typeof response.url !== 'string') {
+  if (typeof response.url !== 'string' || response.url === '') {
     if (typeof input === 'string') {
       response.url = input;
     } else if (input && typeof input.url === 'string') {
@@ -90,7 +91,7 @@ function setMockedResponse(response: MockResponse, input: MockRequestInfo, rejec
     }
   }
 
-  const u = new URL(response.url, FETCH_BASE_URL);
+  const u = new URL(response.url, location.href);
   response.url = u.href;
 
   const mockedResponseData: MockedResponseData = {
@@ -109,18 +110,18 @@ export const mockFetch = {
 
   json(data: any, url?: string) {
     const rsp = new MockResponse(JSON.stringify(data, null, 2), {
-      headers: new MockHeaders([
-        ['Content-Type', 'application/json']
-      ])
+      headers: new MockHeaders({
+        'Content-Type': 'application/json'
+      })
     });
     setMockedResponse(rsp, url, false);
   },
 
   text(data: string, url?: string) {
     const rsp = new MockResponse(data, {
-      headers: new MockHeaders([
-        ['Content-Type', 'text/plain']
-      ])
+      headers: new MockHeaders({
+        'Content-Type': 'text/plain'
+      })
     });
     setMockedResponse(rsp, url, false);
   },
@@ -136,121 +137,15 @@ export const mockFetch = {
   reset: mockFetchReset
 };
 
-export class MockHeaders {
-  values = new Map<string, string>();
-
-  constructor(values?: string[][]) {
-    if (Array.isArray(values)) {
-      values.forEach(v => {
-        this.values.set(v[0], v[1]);
-      });
-    }
-  }
-
-  forEach(cb: (value: string, key: string) => void) {
-    this.values.forEach(cb);
-  }
-
-  serialize() {
-    const s: string[][] = [];
-    this.values.forEach((value, key) => {
-      s.push([key, value]);
-    });
-    return s;
-  }
-
-  append(key: string, value: string) {
-    this.values.set(key, value);
-  }
-
-  delete(key: string) {
-    this.values.delete(key);
-  }
-
-  get(key: string) {
-    return this.values.get(key);
-  }
-
-  has(key: string) {
-    return this.values.has(key);
-  }
-
-  set(key: string, value: string) {
-    this.values.set(key, value);
-  }
-}
-
-
-// ReQUEST
-
-export interface MockRequestInit {
-  headers?: MockHeaders;
-  status?: number;
-  statusText?: string;
-}
-
-export type MockRequestInfo = MockRequest | string;
-
-export class MockRequest {
-  headers?: MockHeaders;
-  method?: 'DELETE' | 'GET' | 'HEAD' | 'OPTIONS' | 'POST' | 'PUT' | 'PATCH';
-  url?: string;
-
-  constructor(_mockInput: MockRequestInfo, _mockInit?: MockRequestInit) {}
-}
-
-
-// ReSPONSE
-
-export interface MockResponseInit {
-  headers?: MockHeaders;
-  ok?: boolean;
-  status?: number;
-  statusText?: string;
-  type?: ResponseType;
-  url?: string;
-}
-
-export class MockResponse {
-  headers?: MockHeaders;
-  ok: boolean;
-  status?: number;
-  statusText?: string;
-  type?: string;
-  url?: string;
-
-  constructor(private mockedBody: string, private mockedResponseInit?: MockResponseInit) {
-    if (mockedResponseInit) {
-      this.headers = mockedResponseInit.headers;
-      this.ok = mockedResponseInit.ok;
-      this.status = mockedResponseInit.status;
-      this.statusText = mockedResponseInit.statusText;
-      this.type = mockedResponseInit.type;
-      this.url = mockedResponseInit.url;
-    }
-  }
-
-  async json() {
-    return JSON.parse(this.mockedBody);
-  }
-
-  async text() {
-    return this.mockedBody;
-  }
-
-  clone() {
-    return new MockResponse(this.mockedBody, this.mockedResponseInit);
-  }
-}
 
 class MockResponse404 extends MockResponse {
   status = 404;
   statusText = 'Not Found'
   constructor() {
     super('', {
-      headers: new MockHeaders([
-        ['Content-Type', 'text/plain']
-      ])
+      headers: new MockHeaders({
+        'Content-Type': 'text/plain'
+      })
     });
   }
   async json() {
@@ -267,6 +162,6 @@ interface MockedResponseData {
   reject: boolean;
 }
 
-const FETCH_BASE_URL = 'https://mocked.stenciljs.com';
-
 const FETCH_DEFAULT_PATH = '/mock-fetch-data';
+
+export { MockHeaders, MockRequest, MockRequestInit, MockRequestInfo, MockResponse, MockResponseInit } from '@stencil/core/mock-doc';