+++ src/testing/platform/testing-host-ref.ts
@@ -13,7 +13,8 @@ export const registerInstance = (lazyInstance: any, hostRef: d.HostRef) => {
   }
 
   if (hostRef == null) {
-    console.warn('Use newSpecPage() to instanciate component instances.');
+    console.warn(`Use "newSpecPage()" to instanciate component instances instead of "new MyComp()".
+For further information: https://stenciljs.com/docs/unit-testing#newspecpage-`);
     const Cstr = lazyInstance.constructor as d.ComponentTestingConstructor;
     const tagName = (Cstr.COMPILER_META && Cstr.COMPILER_META.tagName) ? Cstr.COMPILER_META.tagName : 'div';
     const elm = document.createElement(tagName);